atol

+++ src/ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: flauer <flauer@student.42heilbronn.de>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/16 16:53:40 by flauer            #+#    #+#             */
-/*   Updated: 2023/05/09 10:11:06 by flauer           ###   ########.fr       */
+/*   Updated: 2023/05/09 10:27:51 by flauer           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,11 +22,11 @@ static int	ft_isspace(const char c)
 /// @brief convert a string to an integer
 /// @param str input string with decimal number
 /// @return number as integer
-int	ft_atoi(const char *str)
+long	ft_atoi(const char *str)
 {
-	int	ret;
-	int	i;
-	int	sign;
+	long	ret;
+	int		i;
+	int		sign;
 
 	ret = 0;
 	i = 0;
